machine: Remove dead code
WriteConfig does not appear to have any users. This commit removes it.
WriteConfig was the only user of the VM interface, and it was unused and
removed in f3c82a917c. The VM interface can also be removed.
This was made obsolete by the big refactoring which happened around
commits b1ce6ef, d7cb66492, ...

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>