add support to use cloudInit files passed by user
so far users could not use any custom cloud init config files with podman.
A basic user-data config file was automatically generated by podman to customize ssh access to the new VM through cloud-init.
This patch extends the MachineConfig to store CloudInit config files passed by the user.
If so, podman skips the generation process of the default user-data and uses the user-generated files to customize the VM.

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>
Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>