extend user-data if usermode networking is enabled in hyperv
if usermode networking is enabled on hyperv we need to inject
gvforwarder and set up the vsock connection and a systemd service to
handle it.

To do this we add gvforwarder binary within the ISO and use a custom user-data to copy
it in the right place in the guest.

The resulting user-data is similar to https://github.com/crc-org/macadam/issues/228#issuecomment-3334439056

Signed-off-by: lstocchi <lstocchi@redhat.com>